correct V695 to make it such that nmls ids must be integers and cannot be doubles

+++ hmda/src/main/scala/hmda/validation/rules/lar/validity/2021/V695_1.scala
@@ -12,7 +12,7 @@ object V695_1 extends EditCheck[LoanApplicationRegister] {
   override def parent: String = "V695"
 
   override def apply(lar: LoanApplicationRegister): ValidationResult = {
-    (lar.larIdentifier.NMLSRIdentifier is numeric) or 
+    (lar.larIdentifier.NMLSRIdentifier is integer) or 
     (lar.larIdentifier.NMLSRIdentifier is oneOf("Exempt", "NA"))
   }
